factor out catalog classes
