+++ GCRCatalogs/__init__.py
@@ -9,6 +9,7 @@ from .buzzard import *
 from .dc1 import *
 
 from .gcrCatSimDatabaseEmulator import *
+from .gcrCatSimCatalogClasses import *
 from .gcrCatSimSedFitter import *
 
 from .version import __version__
